requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
absl-py=1.1.0=pypi_0
addict=2.4.0=pypi_0
anyio=3.6.1=pypi_0
anykeystore=0.2=pypi_0
apex=0.1=pypi_0
apted=1.0.3=pypi_0
blas=1.0=mkl
ca-certificates=2022.6.15=ha878542_0
cachetools=5.2.0=pypi_0
certifi=2022.6.15=pypi_0
cffi=1.15.1=pypi_0
charset-normalizer=2.0.12=pypi_0
ci-info=0.3.0=pypi_0
click=8.1.3=pypi_0
cloudpickle=2.1.0=pypi_0
configobj=5.0.6=pypi_0
configparser=5.2.0=pypi_0
cryptacular=1.6.2=pypi_0
cryptography=37.0.4=pypi_0
cudatoolkit=11.0.3=h88f8997_10
cycler=0.11.0=pypi_0
defusedxml=0.7.1=pypi_0
detectron2=0.3+cu110=pypi_0
etelemetry=0.3.0=pypi_0
filelock=3.7.1=pypi_0
fitz=0.0.1.dev2=pypi_0
fonttools=4.33.3=pypi_0
freetype=2.10.4=h0708190_1
future=0.18.2=pypi_0
fvcore=0.1.5.post20220512=pypi_0
google-auth=2.8.0=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
greenlet=1.1.2=pypi_0
grpcio=1.47.0=pypi_0
h11=0.12.0=pypi_0
httpcore=0.15.0=pypi_0
httplib2=0.20.4=pypi_0
httpx=0.23.0=pypi_0
huggingface-hub=0.0.8=pypi_0
hupper=1.10.3=pypi_0
idna=3.3=pypi_0
importlib-metadata=4.11.4=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
iopath=0.1.9=pypi_0
isodate=0.6.1=pypi_0
jarowinkler=1.0.5=pypi_0
joblib=1.2.0=pypi_0
jpeg=9b=h024ee3a_2
kiwisolver=1.4.3=pypi_0
ld_impl_linux-64=2.36.1=hea4e1c9_2
levenshtein=0.18.1=pypi_0
libffi=3.3=h58526e2_2
libgcc-ng=12.1.0=h8d9b700_16
libgomp=12.1.0=h8d9b700_16
libpng=1.6.37=h21135ba_2
libstdcxx-ng=12.1.0=ha89aaad_16
libtiff=4.2.0=h85742a9_0
libuv=1.43.0=h7f98852_0
libwebp-base=1.2.2=h7f98852_1
libzlib=1.2.12=h166bdaf_1
looseversion=1.0.1=pypi_0
lxml=4.9.1=pypi_0
lz4-c=1.9.3=h9c3ff4c_1
markdown=3.3.7=pypi_0
markupsafe=2.1.1=pypi_0
matplotlib=3.5.2=pypi_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py38h95df7f1_0
mkl_fft=1.3.1=py38h8666266_1
mkl_random=1.2.2=py38h1abd341_0
mmcv=1.3.12=pypi_0
mmcv-full=1.3.8=pypi_0
mmdet=2.16.0=pypi_0
ncurses=6.3=h27087fc_1
networkx=2.8.5=pypi_0
nibabel=4.0.1=pypi_0
ninja=1.11.0=h924138e_0
nipype=1.8.3=pypi_0
numpy=1.23.0=pypi_0
numpy-base=1.22.3=py38hf524024_0
oauthlib=3.2.0=pypi_0
olefile=0.46=pyh9f0ad1d_1
opencv-python=4.6.0.66=pypi_0
openssl=1.1.1p=h166bdaf_0
packaging=21.3=pypi_0
pandas=1.4.3=pypi_0
pastedeploy=2.1.1=pypi_0
pathlib=1.0.1=pypi_0
pbkdf2=1.3=pypi_0
pdfminer-six=20220524=pypi_0
pdfplumber=0.7.4=pypi_0
pillow=9.1.1=pypi_0
pip=22.1.2=pyhd8ed1ab_0
plaster=1.0=pypi_0
plaster-pastedeploy=0.7=pypi_0
polygon=1.0.7=pypi_0
polygon3=3.0.9.1=pypi_0
portalocker=2.4.0=pypi_0
prettytable=3.6.0=pypi_0
protobuf=3.19.4=pypi_0
prov=2.0.0=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycocotools=2.0.4=pypi_0
pycparser=2.21=pypi_0
pydot=1.4.2=pypi_0
pyparsing=3.0.9=pypi_0
pyramid=2.0=pypi_0
pyramid-mailer=0.15.1=pypi_0
python=3.8.8=hffdb5ce_0_cpython
python-dateutil=2.8.2=pypi_0
python3-openid=3.2.0=pypi_0
python_abi=3.8=2_cp38
pytorch=1.7.1=py3.8_cuda11.0.221_cudnn8.0.5_0
pytz=2022.1=pypi_0
pyxnat=1.4=pypi_0
pyyaml=6.0=pypi_0
rapidfuzz=2.1.1=pypi_0
rdflib=6.2.0=pypi_0
readline=8.1.2=h0f457ee_0
regex=2022.6.2=pypi_0
repoze-sendmail=4.4.1=pypi_0
requests=2.28.0=pypi_0
requests-oauthlib=1.3.1=pypi_0
rfc3986=1.5.0=pypi_0
rsa=4.8=pypi_0
sacremoses=0.0.53=pypi_0
scikit-learn=1.2.1=pypi_0
scipy=1.9.0=pypi_0
seaborn=0.12.2=pypi_0
sentencepiece=0.1.97=pypi_0
setuptools=62.6.0=py38h578d9bd_0
simplejson=3.17.6=pypi_0
six=1.16.0=pyh6c4a22f_0
sklearn=0.0.post1=pypi_0
sniffio=1.2.0=pypi_0
sqlalchemy=1.4.39=pypi_0
sqlite=3.38.5=h4ff8645_0
tabulate=0.8.10=pypi_0
tensorboard=2.9.1=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
termcolor=1.1.0=pypi_0
terminaltables=3.1.10=pypi_0
threadpoolctl=3.1.0=pypi_0
tk=8.6.12=h27826a3_0
tokenizers=0.10.3=pypi_0
torch=1.7.1=pypi_0
torchaudio=0.7.2=pypi_0
torchvision=0.8.2=pypi_0
tqdm=4.64.0=pypi_0
traits=6.3.2=pypi_0
transaction=3.0.1=pypi_0
transformers=4.6.0=pypi_0
translationstring=1.4=pypi_0
typing_extensions=4.2.0=pyha770c72_1
urllib3=1.26.9=pypi_0
velruse=1.1.1=pypi_0
venusian=3.0.0=pypi_0
wand=0.6.8=pypi_0
webob=1.8.7=pypi_0
websocket-client=1.3.3=pypi_0
websockets=10.3=pypi_0
werkzeug=2.1.2=pypi_0
wheel=0.37.1=pyhd8ed1ab_0
wtforms=3.0.1=pypi_0
wtforms-recaptcha=0.3.2=pypi_0
xz=5.2.5=h516909a_1
yacs=0.1.8=pypi_0
yapf=0.32.0=pypi_0
zipp=3.8.0=pypi_0
zlib=1.2.12=h166bdaf_1
zope-deprecation=4.4.0=pypi_0
zope-interface=5.4.0=pypi_0
zope-sqlalchemy=1.6=pypi_0
zstd=1.4.9=ha95c52a_0
